refactor : split Test Asserts for each page in EasyAdmin
