+++ src/Test/Trait/CrudTestFormAsserts.php
@@ -1,0 +1,48 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EasyCorp\Bundle\EasyAdminBundle\Test\Trait;
+
+trait CrudTestFormAsserts
+{
+    protected function assertFormFieldExists(string $fieldName): void
+    {
+        // TODO : to implement
+    }
+
+    protected function assertFormFieldNotExists(string $fieldName): void
+    {
+        // TODO : to implement
+    }
+
+    protected function assertFormFieldHasLabel(string $fieldName, string $label): void
+    {
+        // TODO : to implement
+    }
+
+    protected function assertFormFieldNotHasLabel(string $fieldName, string $label): void
+    {
+        // TODO : to implement
+    }
+
+    protected function assertFormFieldHasValue(string $fieldName, string|int|bool $value): void
+    {
+        // TODO : to implement
+    }
+
+    protected function assertFormFieldNotHasValue(string $fieldName, string|int|bool $value): void
+    {
+        // TODO : to implement
+    }
+
+    protected function assertFormFieldIsDisplayed(string $fieldName): void
+    {
+        // TODO : to implement
+    }
+
+    protected function assertFormFieldIsHidden(string $fieldName): void
+    {
+        // TODO : to implement
+    }
+}
